backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlite3
python-multipart==0.0.6
groq==0.4.2
opencv-python==4.0.1.78
numpy==1.24.3
tensorflow-cpu==2.13.0
pydantic==2.5.0
python-dotenv==1.0.0